!Change Button Status!

+++ Rap/src/database/LinkConnector.java
@@ -217,7 +217,7 @@ public class LinkConnector {
 	 */
 	public static int getWorkersCount() {
 		Query q = entityManager.createQuery("SELECT COUNT(1) FROM Worker w");
-		return (Integer) q.getSingleResult();
+		return ((Long) q.getSingleResult()).intValue();
 	}
 	
 	/**
